% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/engineerbunny-package.R
\docType{package}
\name{engineerbunny-package}
\alias{engineerbunny}
\alias{engineerbunny-package}
\title{engineerbunny: Data Engineering Tools}
\description{
Engineer datasets (just how I like it).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://danielbraddock.github.io/engineerbunny/}
}

}
\author{
\strong{Maintainer}: Daniel Braddock \email{nobody@noreply.com}

}
\keyword{internal}
