% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keys_check.R
\name{keys_check}
\alias{keys_check}
\title{Check columns form a key}
\usage{
keys_check(df, keys)
}
\arguments{
\item{df}{A data frame, data frame extension (e.g. a tibble), or a
lazy data frame (e.g. from dbplyr or dtplyr).}

\item{keys}{Selection of columns using tidyselect}
}
\value{
Whether the given variables form a key
ie are sufficient to specify uniqueness and completeness. TRUE / FALSE
}
\description{
Indicates whether the columns selected as keys truly do form a key

ie are sufficient to identify unique and complete rows
}
\details{
When TRUE it is still possible for a sub-selection to form a key
}
\examples{
\dontrun{
iris |> mutate(Species_id = row_number(), .by = Species) |>
keys_check(contains("Species")) # TRUE
iris |> mutate(id = row_number()) |> keys_check(id) # TRUE
iris |> keys_check(everything()) # FALSE
iris |> distinct() |> keys_check(everything()) # TRUE
iris |> keys_check(c(starts_with("Sepal"), starts_with("Petal"))) # FALSE
iris |> keys_check(starts_with("Sepal")) # FALSE
iris |> keys_check(Species) # FALSE
}
}
