% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keys_count.R
\name{keys_count}
\alias{keys_count}
\title{Count the times each unique key level appears and hence number of dupes}
\usage{
keys_count(df, keys)
}
\arguments{
\item{df}{A data frame, data frame extension (e.g. a tibble), or a
lazy data frame (e.g. from dbplyr or dtplyr).}

\item{keys}{Selection of columns using tidyselect}
}
\value{
A data.frame, or dtplyr_step_mutate, of results listing the unique levels of the keys and counts
}
\description{
Count the times each unique key level appears and hence number of dupes
}
\examples{
\dontrun{
iris |> mutate(Species_id = row_number(), .by = Species) |> keys_count(Species_id)
iris |> mutate(id = row_number()) |> keys_count(id)
}
}
