% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keys_interrogate_failure.R
\name{keys_interrogate_failure}
\alias{keys_interrogate_failure}
\title{Find and compare an example of duplication}
\usage{
keys_interrogate_failure(df, keys, dupe_severity = 1, dupe_id = 1)
}
\arguments{
\item{df}{A data frame, data frame extension (e.g. a tibble), or a
lazy data frame (e.g. from dbplyr or dtplyr).}

\item{keys}{Selection of columns using:
* tidyselect
* names (string) ie a character vector

Not like the other keys_*() functions, not using:
* unquoted expressions}

\item{dupe_severity}{The level of duplication required for the comparison}

\item{dupe_id}{Each example is given an id. Change this for a different example.}
}
\value{
A data.frame listing the columns which ideally would form a key and
the columns which are responsible for that key not being a key after all
}
\description{
Find and compare an example of duplication
}
\examples{
\dontrun{
iris |> mutate(x = row_number()) |> keys_interrogate_failure(starts_with("Sepal"))
iris |> mutate(x = row_number()) |> keys_interrogate_failure(c("Sepal.Length"
, "Sepal.Width", "Petal.Length"))
iris |> mutate(x = row_number()) |> keys_interrogate_failure(c("Sepal.Length"
, "Sepal.Width", "Petal.Length", "Petal.Width"))
try(iris |> mutate(x = row_number()) |> keys_interrogate_failure(c("Sepal.Length"
, "Sepal.Width", "Petal.Length", "Petal.Width", "x")))
}
}
