% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_feather.R
\name{scrape_feather}
\alias{scrape_feather}
\title{Read all feather files after a pattern into a list of datasets}
\usage{
scrape_feather(dir, as_data_table = TRUE, pattern = ".*", negate = FALSE)
}
\arguments{
\item{dir}{<chr> file path to scrape for .feather files}

\item{as_data_table}{whether to convert the dataset to data.table}

\item{pattern}{<chr> regex to match in file names}

\item{negate}{<chr> passed to str_detect when selecting files}
}
\value{
named list of datasets scraped from dir, named to match file name, defaults to TRUE
}
\description{
Read all feather files after a pattern into a list of datasets
}
\examples{
\dontrun{splits <- scrape_feather("Data")}
}
